refactor: establish internal markdown diagram transform
